fix build.yml

+++ src/csv/CsvParser.h
@@ -15,7 +15,7 @@ class CsvParser
 {
     void ValidateValue(const std::string value);
 
-    int _lineCounter = 0;
+    size_t _lineCounter = 0;
     fs::path _file = {};
     std::ifstream _ifstream = {};
     CsvFormat _format = {};
